Move removeSessionCookie to sessions/http/
