+++ packages/core/src/sessions/http/remove-session-cookie.spec.ts
@@ -1,5 +1,8 @@
+// std
 import { strictEqual } from 'assert';
-import { Config, HttpResponse, HttpResponseOK } from '../core';
+
+// FoalTS
+import { Config, HttpResponse, HttpResponseOK } from '../../core';
 import {
   SESSION_DEFAULT_COOKIE_HTTP_ONLY,
   SESSION_DEFAULT_COOKIE_NAME,
@@ -7,7 +10,7 @@ import {
   SESSION_DEFAULT_CSRF_COOKIE_NAME,
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
-} from './constants';
+} from '../constants';
 import { removeSessionCookie } from './remove-session-cookie';
 
 describe('removeSessionCookie', () => {
