+++ packages/core/src/sessions/use-sessions.hook.ts
@@ -22,7 +22,7 @@ import { createSession } from './create-session';
 import { checkUserIdType } from './http/check-user-id-type';
 import { getSessionIDFromRequest, RequestValidationError } from './http/get-session-id-from-request';
 import { readSession } from './read-session';
-import { removeSessionCookie } from './remove-session-cookie';
+import { removeSessionCookie } from './http/remove-session-cookie';
 import { SessionStore } from './session-store';
 import { setSessionCookie } from './set-session-cookie';
